Separate data copying and loading into two distinct phases
This commit changes the tool so that copying the data from the sources databases
and loading it into the target MonetDB are two distinct separate phases.

First, all the data will be copied to the local disk and then the data will be loaded into
the target MonetDB database. This has the added potential benefit that data may be
directly loaded into MonetDB without having to stream it using STDIN.