Make fast view-switching more efficient
This commit will make the fast view-switching functionality much more
efficient by using a date postfix on the table names instead of the
backup/current prefix. By doing this no data will have to be copied to
the backup table but instead new data is loaded into a newer table.